vefvzrbe

+++ PHP/deleteSeance.php
@@ -1,6 +1,13 @@
 <?php
 
-require_once 'DB/DB.inc.php';
+chdir("..");
+
+session_start();
+require_once "PHP/fonctions/func.seance.php";
+require_once("PHP/Twig/lib/Twig/Autoloader.php");
+require_once "PHP/fonctions/func.typeseance.php";
+require_once "PHP/fonctions/func.modules.php";
+require_once "PHP/fonctions/func.groupes.php";
 
 
 // TODO: à transformer en fonction
@@ -10,17 +17,26 @@ $db = DB::getInstance();
 if ($db == null) {
     echo "Impossible de se connecter &agrave; la base de donneeacute;es !";
 } else {
-    try {
-		$db->deleteEvenementsForSeance($_GET['id']);
-		$db->deleteSemaphoresForSeance($_GET['id']);
-		$db->deleteSeance($_GET['id']);
-        echo '<meta http-equiv="refresh" content="0;URL=../listEvenement.php">';
+	if(isset($_GET['id'])) {
 
-    } catch (Exception $e) {
-        echo "ça MARCHE PAS: " . $e->getMessage();
+		echo 'test2';
+    	$seance = getSeanceIdSc($_GET['id'])[0];
+    	echo $seance;
+    	if( $seance->getIdUtilisateur() === $_SESSION['login']) {
+    		echo "c'est les memes personnes";
+    		try {
+				$db->deleteEvenementsForSeance($_GET['id']);
+				$db->deleteSemaphoresForSeance($_GET['id']);
+				$db->deleteSeance($_GET['id']);
+    		    echo '<meta http-equiv="refresh" content="0;URL=../listSeance.php">';
+	
+	    	} catch (Exception $e) {
+	    		echo "ça MARCHE PAS: " . $e->getMessage();
+		
+    		}
+		}
+	}
+	echo '<meta http-equiv="refresh" content="0;URL=../listSeance.php">';
 
-    }
 }
-
-
 $db->close();
